refactor: Image 패키지 이동
